# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GitHub Integration API",
        "description": "API for interacting with GitHub repositories",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:1212",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/repositories": {
            "get": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Get list of repositories",
                "description": "Returns a paginated list of repositories for the authenticated user",
                "operationId": "getRepositories",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Field to sort by (created, updated, pushed, full_name)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "updated"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Repository"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/repository": {
            "get": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Get specific repository details",
                "description": "Returns details of a specific repository",
                "operationId": "getRepository",
                "parameters": [
                    {
                        "name": "owner",
                        "in": "query",
                        "description": "Repository owner/organization name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repo",
                        "in": "query",
                        "description": "Repository name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Repository"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Repository not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/clear-repository-cache": {
            "post": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Clear repository cache",
                "description": "Clears the cached repository data for the current user",
                "operationId": "clearRepositoryCache",
                "responses": {
                    "200": {
                        "description": "Cache cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Repository cache cleared successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/pull-repository-code": {
            "post": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Pull repository code",
                "description": "Pulls code from a specific repository",
                "operationId": "pullRepositoryCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "owner",
                                    "repo"
                                ],
                                "properties": {
                                    "owner": {
                                        "type": "string",
                                        "example": "octocat"
                                    },
                                    "repo": {
                                        "type": "string",
                                        "example": "hello-world"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code pulled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Repository code pulled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Repository not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Repository": {
                "title": "Repository",
                "description": "GitHub repository model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123456789
                    },
                    "name": {
                        "type": "string",
                        "example": "repo-name"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "octocat/repo-name"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a sample repository",
                        "nullable": true
                    },
                    "private": {
                        "type": "boolean",
                        "example": false
                    },
                    "html_url": {
                        "type": "string",
                        "example": "https://github.com/octocat/repo-name"
                    },
                    "language": {
                        "type": "string",
                        "example": "PHP",
                        "nullable": true
                    },
                    "stargazers_count": {
                        "type": "integer",
                        "example": 42
                    },
                    "watchers_count": {
                        "type": "integer",
                        "example": 42
                    },
                    "forks_count": {
                        "type": "integer",
                        "example": 23
                    },
                    "open_issues_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "default_branch": {
                        "type": "string",
                        "example": "main"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-02T00:00:00Z"
                    },
                    "pushed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-03T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}